Deleting artwork

+++ src/util/index.js
@@ -75,28 +75,22 @@ const putData = async (url, data, config = {}) => {
 };
 
 const deleteData = async (url, params = {}, headers = {}) => {
-  const token = localStorage.getItem("token");
-  const config = {
-    params,
-    headers: {
-      'Authorization': `Bearer ${token}`,
-      ...headers,
-    },
-  }; const res = await axios.delete(url, config);
-  const data = res.data;
-  return data;
+  const user = JSON.parse(localStorage.getItem("user"))
+  if (user) {
+    const config = {
+      params,
+      headers: {
+        'Authorization': `Bearer ${user.token}`,
+        ...headers,
+      },
+    }; const res = await axios.delete(url, config);
+    const data = res.data;
+    return data;
+  } else {
+    throw new Error("User not logged in");
+  }
 };
-// try {
-//   const res = await axios.delete(url, {
-//     ...config,
-//     withCredentials: true,
-//   });
-//   return res.data;
-// } catch (error) {
-//   console.log(error, `error - deleteData in ${url} route`);
-//   throw error;
-// }
-// };
+
 
 const isEmpty = (value) => {
   return (
